% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_prospect.R
\name{fields_prospect}
\alias{fields_prospect}
\title{Field Descriptions}
\usage{
fields_prospect(df = master$screening, fields = fields)
}
\arguments{
\item{df}{The data frame who's variables need describing}

\item{fields}{A data frame that contains the information from the database specification
spreadsheet produced by Data Management. It should have a bare minimum of two
columns the 'Identifier' column (normally column D) and the 'Label' column
(usually column E) and these should be named \code{variable} and
\code{description} respectively.}
}
\value{
Data frame of two columns, \code{variable} listing the variables in the data frame
        and \code{description} which provides the description of the variable.
}
\description{
Descriptions of fields (variables) in the Dipep dataset
}
\details{
Variable names are sometimes informative, sometimes cryptic.  To avoid this each data frames
variables are described using the database specification produced by Data Management.
}
\examples{

## You will need to download the 'fields' form from the Data Management database
## specification, read it in and rename columns appropriately.
fields <- <- read.csv('fields.csv')
fields <- dplyr::select(fields, Identifier, Label)
names(fields) <- c('variable', 'description')
## Finally because I work with '.' rather than '_' between words convert...
fields$variable <- gsub('_', '.', fields$variable)
## Save a description of your variables to a data frame that is part of alist of
## data frames (since there are multiple data frames in a given project)
README.variables <- list()
README.variables$screening <- fields_prospect(df     = master$screening,
                                              fields = fields)


}

